Add files via upload
This is an example of a simple n-gram language model implementation, using only the random library. Note for beginners: download the accompanying .txt file and try prompting for the kind of information you would expect the model to give you if it was trained on the example Wikipedia page; otherwise collect the training data yourself and build upon the code further.